BREAKING CHANGE: no more flags, autodetect running daemon, start otherwise
Now, if -k flag is used, the daemon subcommand checks if there is a
daemon running on the socket, if so, do not start a new one.
All other subcommands now also do this check, and if they are not
responded to in 160ms then the socket will be deleted and a new daeomn
process will be spawned.